+++ python/tests/test_network.py
@@ -2137,7 +2137,7 @@ class TestNetwork(unittest.TestCase):
                 dP = np.max([gen.P-gen.P_max,gen.P_min-gen.P,0.])
                 if dP > Pvio:
                     Pvio = dP
-                if gen.is_regulator():
+                if gen.is_regulator() and not gen.is_slack():
                     dQ = np.max([gen.Q-gen.Q_max,gen.Q_min-gen.Q,0.])
                     if dQ > Qvio:
                         Qvio = dQ
@@ -4358,28 +4358,34 @@ class TestNetwork(unittest.TestCase):
         
         for case in test_cases.CASES:
 
+            import os
+            if os.path.basename(case) in ['sys_problem2.mat', 'sys_problem3.mat']:
+                continue
+
             net = pf.Parser(case).parse(case, num_periods=2)
 
-            net.set_flags('generator',
-                          'variable',
-                          'regulator',
-                          'reactive power')
-
-            for gen in net.generators:
-                if gen.is_regulator():
-                    gen.Q_min = 0.
-                    gen.Q_max = 0.
+            net0 = net.get_copy()
+            net0.update_properties()
 
             net.adjust_generators()
 
-            x = net.get_var_values()
-            self.assertFalse(np.any(np.isnan(x)))
-            
-            constr = pf.Constraint('generator reactive power participation', net)
-            constr.analyze()
-            A = constr.A
-            b = constr.b
-            self.assertLess(np.linalg.norm(A*x-b),1e-12)
+            net1 = net.get_copy()
+            net1.update_properties()
+
+            for i in range(net.num_buses):
+                for t in range(net.num_periods):
+                    bus0 = net0.get_bus(i)
+                    bus1 = net1.get_bus(i)
+                    if not [g for g in bus0.generators if g.is_slack()]:
+                        self.assertEqual(bus0.P_mismatch[t], bus1.P_mismatch[t])
+                    else:
+                        self.assertNotEqual(bus0.P_mismatch[t], bus1.P_mismatch[t])
+                        self.assertLess(np.abs(bus1.P_mismatch[t]), 1e-10)
+                    if not [g for g in bus0.generators if g.is_regulator()]:
+                        self.assertEqual(bus0.Q_mismatch[t], bus1.Q_mismatch[t])
+                    else:
+                        self.assertNotEqual(bus0.Q_mismatch[t], bus1.Q_mismatch[t])
+                        self.assertLess(np.abs(bus1.Q_mismatch[t]), 1e-10)                    
 
     def test_symmetric_connectors_removers(self):