+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Unreleased
 * Added Q_par (reactive power participation) field to generator.
 * Added "type" parameter to PAR_GEN_Q constraint that allows gen Q participations to be based on Q_par gen fields.
 * Changed REG_GEN constraint to treat generators separately so that constraint can be used without participations.
+* Updated network "adjust generators" to balance bus mismatches with slack and reg generators.
 
 Version 1.3.2
 -------------